remove level & move did-web server to tests
