+++ packages/did-resolver/src/index.ts
@@ -1,4 +1,4 @@
-export * as web from './web/resolver'
-export * as plc from './plc/resolver'
+export * as web from './web-resolver'
+export * as plc from './plc-resolver'
 export * from './resolver'
 export * from './atp-did'
