+++ packages/plc/build.js
@@ -29,8 +29,6 @@ require('esbuild').build({
   plugins: buildShallow ? [nodeExternalsPlugin()] : [],
   external: [
     'better-sqlite3',
-    'level',
-    'classic-level',
     // Referenced in pg driver, but optional and we don't use it
     'pg-native',
   ],
